option_parser: Move test_option_parser to option_parser crate
